# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Result": {
				"properties": {
					"link": {
						"type": "string"
					}
				},
				"required": [
					"link"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Error": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "kawaii-image-api",
		"version": "0.1",
		"description": "REST API to serve different images or gifs from animes",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "eme22"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/sfw/kiss": {
			"get": {
				"operationId": "GetRandomKiss",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Theres no images in this endpoint",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves a random Kissing GIF link",
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": []
			}
		},
		"/sfw/kiss/{id}": {
			"get": {
				"operationId": "GetKiss",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Result"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Image not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"tags": [
					"SFW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}